Allow RestartGameplay When Mid-Fail
The old behavior allowed for the player to
hit the restart button in the middle of
failing or at the end of a file.
This makes that possible again, since the
previous change forced the player to only
be allowed to hit RestartGameplay when in
the middle of a file, not transitioning.