Add Support for Pre-v.60 Replays
Also fix HighScore mysteriously changing old loaded replays to appear like new ones..?
This utilizes the int created in HighScore that keeps track of the replay type.
Since we have no data on holds, mines, or columns in the old replay data we just go ahead and hit every note that we know was hit. If more than one are on a line, then it biases left and places judgements however it wants. The overall judgement should be correct.
Visually (unless implemented in Lua) there is no indication that a Replay is old or not.
An old and new replay appear exactly the same but may score slightly differently due to holds and mines.