Add Top Score/All Scores toggle like EO
Modified the construction of the score leaderboards in DLMAN. Before, it skipped duplicate names always. Now, it always retrieves the whole leaderboard.
The method which retrives the leaderboard dummies for Lua has been modified to cope with this.
It seems this is the only reason we would want to filter out duplicate scores anyways.
New lua functions have been added to modify and retrieve the boolean associated with the changes.